services/gaming_billing/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.0
django-debug-toolbar==5.1.0
django-filter==25.1
django-ipware==7.0.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
docopt==0.6.2
factory_boy==3.3.3
Faker==37.1.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.3
packaging==24.2
pip-upgrader==1.4.15
prompt_toolkit==3.0.51
psycopg==3.2.6
psycopg-binary==3.2.6
python-crontab==3.2.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
toml==0.10.2
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
